kafel: Add SECCOMP_RET_LOG support
SECCOMP_RET_LOG is a new seccomp action meant
to aid development and debugging.  This action
will prompt the kernel to log the syscall
violation, but continue executing the syscall
as if nothing happened.  This allows you to log
all violated syscalls (instead of doing the classic
add-one-syscall-at-a-time approach).

To use this action in Kafel, simply use the LOG
action as the default, i.e.

USE <policy> DEFAULT LOG

Note that if your kernel does not support this
action, your filter will instead default to
SECCOMP_RET_KILL.