Add bitwise OR, multiple bit ops
This change allows the bitwise or (`|`) operator to be used in argument
filters. E.g.:

```
open { flags == O_RDONLY|O_CLOEXEC }
```

It also allows multiple uses of both AND (`&`) and OR within an
expresssion:

```
open { SAFE_FLAGS2 & flags & SAFE_FLAGS1 == flags }
```

The number of bitwise operations that can occur in a single expression
is limited by the available BPF scratch memory space.